Extract sortItems to utils
